python_checkvar/requirements.txt
asgiref==3.8.1
Authlib==1.3.2
blinker==1.8.2
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
distro==1.9.0
Django==2.1.15
django-mssql-backend==2.8.1
django-pyodbc-azure==2.1.0.0
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Mail==0.10.0
fonttools==4.53.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mysql-connector-python==9.0.0
networkx==3.3
numpy==2.1.0
oauthlib==2.1.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pycparser==2.22
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyodbc==5.1.0
pyparsing==3.1.4
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
requests==2.32.3
requests-oauthlib==1.1.0
scipy==1.14.1
seaborn==0.13.2
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
tabula==1.0.5
tabula-py==2.9.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
